MP Config - enable CDI extension only if Config container deployed
Optimization - MP Config should be completely disabled if not used by the application
