--- appserver/microprofile/config/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-io.helidon.microprofile.config.ConfigCdiExtension
